Import a dataset to an MCView project

Read an anndata file which is the output of python metacells package, and import the metacell dataset to MCView. Each project can have multiple datasets which can be in the app using the right sidebar.

import_dataset(
  project,
  dataset,
  anndata_file,
  cell_type_field = "cluster",
  metacell_types_file = NULL,
  cell_type_colors_file = NULL,
  metadata_fields = NULL,
  metadata = NULL,
  metadata_colors = NULL,
  calc_gg_cor = TRUE
)

Arguments

project

path to the project

dataset

name for the dataset, e.g. "PBMC"

anndata_file

path to h5ad file which contains the output of metacell2 pipeline (metacells python package).

cell_type_field

name of a field in the anndata object$obs which contains a cell type (optional). If the field doesn't exist and metacell_types_file are missing, MCView would cluster the metacell matrix using kmeans++ algorithm (from the tglkmeans package). If metacell_types parameter is set this field is ignored.

metacell_types_file

path to a tabular file (csv,tsv) with cell type assignement for each metacell. The file should have a column named "metacell" with the metacell ids and another column named "cell_type" or "cluster" with the cell type assignment. Metacell ids that do not exists in the data would be ignored. If this parameter and cell_type_field are missing, MCView would cluster the metacell matrix using kmeans++ algorithm (from the tglkmeans package).

cell_type_colors_file

path to a tabular file (csv,tsv) with color assignement for each cell type. The file should have a column named "cell_type" or "cluster" with the cell types and another column named "color" with the color assignment. Cell types that do not exist in the metacell types would be ignored. If this is missing, MCView would use the chameleon package to assign a color for each cell type.

metadata_fields

names of fields in the anndata object$obs which contains metadata for each metacell. The fields should *always* be numeric - if you have cell categorical annotations use cell_metadata_to_metacell with categorical=TRUE to convert them to a numeric score (e.g. using fraction of the category). You can use 'all' in order to import all the numeric fields of the anndata object.

metadata

can be either a data frame with a column named "metacell" with the metacell id and other metadata columns or a name of a delimited file which contains such data frame.

metadata_colors

a named list with colors for each metadata column, or a name of a yaml file with such list. Colors should be given as a list where the first element is a vector of colors and the second element is a vector of breaks. If only colors are given breaks would be implicitly determined from the minimum and maximum of the metadata field.

calc_gg_cor

calculate top 30 correlated and anti-correlated genes for each gene. This computation can be heavy for large datasets or weaker machines, so you can set calc_gg_cor=FALSE to skip it. Note that then this feature would be missing from the app.

Details

The function would create a directory under project/cache/dataset which would contain objects used by MCView shiny app (such as the metacell matrix).

In addition, you can supply file with type assignment for each metacell (metacell_types_file) and a file with color assignment for each metacell type (cell_type_colors_file).

Examples

if (FALSE) {
dir.create("raw")
download.file(
    "http://www.wisdom.weizmann.ac.il/~atanay/metac_data/PBMC_processed.tar.gz",
    "raw/PBMC_processed.tar.gz"
)
untar("raw/PBMC_processed.tar.gz", exdir = "raw")
create_project("PBMC")
import_dataset("PBMC", "PBMC163k", "raw/metacells.h5ad")
}