Import a dataset to an MCView project

This would read would read an anndata file which is the output of python metacells package, and import the metacell dataset to MCView. The result would be a directory under project/cache/dataset which would contain objects used by MCView shiny app (such as the metacell matrix). In addition, you can supply file with type assignment for each metacell (metacell_types_file) and a file with color assignment for each metacell type (cell_type_colors_file).

import_dataset(
  project,
  dataset,
  anndata_file,
  cell_type_field = NULL,
  metacell_types_file = NULL,
  cell_type_colors_file = NULL
)

Arguments

project
path to the project

dataset
name for the dataset, e.g. "PBMC"

anndata_file
path to h5ad file which contains the output of metacell2 pipeline (metacells python package).

cell_type_field
name of a field in the anndata object$obs which contains a cell type (optional). If this parameter and metacell_types_file are missing, MCView would cluster the metacell matrix using kmeans++ algorithm (from the tglkmeans package). If metacell_types parameter is set this field is ignored.

metacell_types_file
path to a tabular file (csv,tsv) with cell type assignement for each metacell. The file should have a column named "metacell" with the metacell ids and another column named "cell_type" or "cluster" with the cell type assignment. Metacell ids that do not exists in the data would be ignored. If this parameter and cell_type_field are missing, MCView would cluster the metacell matrix using kmeans++ algorithm (from the tglkmeans package).

cell_type_colors_file
path to a tabular file (csv,tsv) with color assignement for each cell type. The file should have a column named "cell_type" or "cluster" with the cell types and another column named "color" with the color assignment. Cell types that do not exist in the metacell types would be ignored. If this is missing, MCView would use the chameleon package to assign a color for each cell type.

Examples

if (FALSE) {
dir.create("raw")
download.file("http://www.wisdom.weizmann.ac.il/~atanay/metac_data/PBMC_processed.tar.gz", "raw/#' PBMC_processed.tar.gz")
untar("raw/PBMC_processed.tar.gz", exdir = "raw")
create_project("PBMC")
import_dataset("PBMC", "PBMC163k", "raw/metacells.h5ad")
}